fix: align test helpers with the unified Rect/DisplayId types
Follow-up to reconciling with the already-merged domain model: the
engine's test-only Display builder and the event-hooks test still
referenced the pre-rebase DisplayId(u64) and the now-removed Bounds
type.

+++ crates/mosaix-engine/src/lib.rs
@@ -191,7 +191,7 @@ mod tests {
     use mosaix_domain::{DisplayId, Rect, Rotation};
     use std::time::{Duration, Instant};
 
-    fn display(id: u64, fingerprint: &str, x: i32) -> Display {
+    fn display(id: isize, fingerprint: &str, x: i32) -> Display {
         Display {
             id: DisplayId(id),
             stable_fingerprint: fingerprint.to_string(),
